feat: photoAlBumViewController 레이아웃 적용 requestPhotoAlbum 기능 없애고 다시 AppDelegate 에서 동작하게 코드 수정
