+++ OpenMarket/OpenMarket/AppDelegate.swift
@@ -5,6 +5,9 @@
 // 
 
 import UIKit
+import Photos
+
+var allPhotos: PHFetchResult<PHAsset>?
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -13,24 +16,40 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             return URLCache(memoryCapacity: 500*1024*1024,
                             diskCapacity: 500*1024*1024)
         }()
+        
+        PHPhotoLibrary.requestAuthorization { (status) in
+            switch status {
+            case .authorized:
+                print("Good to proceed")
+                let fetchOptions = PHFetchOptions()
+                allPhotos = PHAsset.fetchAssets(with: .image, options: fetchOptions)
+            case .denied, .restricted:
+                print("Not allowed")
+            case .notDetermined:
+                print("Not determined yet")
+            default:
+                print("error")
+            }
+        }
+        
         // Override point for customization after application launch.
         return true
     }
-
+    
     // MARK: UISceneSession Lifecycle
-
+    
     func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
         // Called when a new scene session is being created.
         // Use this method to select a configuration to create the new scene with.
         return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
     }
-
+    
     func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
         // Called when the user discards a scene session.
         // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
-
-
+    
+    
 }
